chore(prettier): update prettier config and plugins
- Update `.prettierrc.json`:
  - Modify `importOrder` to handle React/Next imports before third-party and local imports
  - Add `prettier-plugin-tailwindcss` plugin for better formatting of Tailwind CSS classes

- Update `package.json`:
  - Add `prettier-plugin-tailwindcss` as a dev dependency for better Tailwind CSS formatting

This change updates the Prettier configuration and dependencies to improve the formatting of React/Next and Tailwind CSS code within the project. The updated import order in `.prettierrc.json` ensures that React/Next imports are grouped together and sorted before other imports, while the addition of the `prettier-plugin-tailwindcss` plugin helps maintain a consistent and readable formatting for Tailwind CSS classes.